requirements.txt
discord.py==1.3.3
colorama==0.4.3
